Deleting Page Object Model Directory
This directory is being deleted to make room for the previous commit titled "Completed Smoke Test." Since the POM class names were changed for clarity purposes they were bunched in with the old POM classes. This delete will be followed by a re-upload of the new Page Object Model Directory.