Refactoring
This is part of the refactoring commit. All commits with the title, "Refactoring", are part of the same update. This update changed namespaces, directory structures, and class names in order to improve workflow/clarity. This update also includes new classes for every page that will be tested in the future. These new classes are the default classes created in C#. They will be updated in a future commit.